fixed inverted blocks, added tinting that applied when r == g == b, made vertices smaller by using float16 for normals and positions. Next to do is implement Face UVS

+++ src/client/java/me/zombii/beryllium/client/rendering/TintProvider.java
@@ -17,7 +17,7 @@ public class TintProvider {
     }
 
     public interface TintFunction {
-        int getTint(BlockState state, BlockPosition pos, int tintIdx);
+        short getTint(BlockState state, BlockPosition pos, int tintIdx);
     }
 
 }
